+++ src/signatures.rs
@@ -395,6 +395,31 @@ signatures! {
     value = b"-lz5-" offset = 2
     value = b"-lhd-" offset = 2
 
+    format = MachO
+    value = b"\xCA\xFE\xBA\xBE", b"\x01" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x02" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x03" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x04" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x05" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x06" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x07" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x08" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x09" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x0A" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x0B" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x0C" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x0D" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x0E" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x0F" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x10" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x11" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\x12" offset = 7
+    value = b"\xCA\xFE\xBA\xBE", b"\xFF" offset = 7
+    value = b"\xFE\xED\xFA\xCE"
+    value = b"\xFE\xED\xFA\xCF"
+    value = b"\xCE\xFA\xED\xFE"
+    value = b"\xCF\xFA\xED\xFE"
+
     format = Opentype
     value = b"\x4F\x54\x54\x4F\x00"